Move ParquetMultiFileInfo to a separate file
